requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
Pillow>=10.0.0
tqdm>=4.66.0
gradio>=4.44.0
onnx>=1.16.0
wandb>=0.17.0
scipy>=1.12.0
